addition number 1 to version 0.1
I added the get_way function to the file_explorer_function.py file, which gives the saved path to the corresponding user. It was made for the convenience of the user so that he writes only the name of the file and not the entire path to it (although I left this option). Added the ability to show file content in text format.